Vertical Slides - single encoder based running
Changed to use only one encoder so that the PIDs of both motors dont conflict. The solution is to keep the motors in sync by using the power reading from the one with the encoder on the one without the encoder. Since the direction of one of them is reversed, the power value *should* effectively translate laterally. Also added a function within the opmode to always keep the motors in sync.